+++ tests/seams/test_tier2_flow.py
@@ -45,7 +45,6 @@ ingress:
   chat_tab: true
   signal: false
   slack: false
-  openwebui: false
 """